decision_agent.py
class DecisionAgent:
    def __init__(self, data_sources):
        self.data_sources = data_sources
        self.models = {}
        
    def analyze_data(self, context):
        """Analyze data from multiple sources"""
        insights = []
        for source in self.data_sources:
            data = source.fetch_data()
            processed_data = self.preprocess(data)
            insight = self.generate_insight(processed_data, context)
            insights.append(insight)
        return self.synthesize_insights(insights) 